edited changes in auth_handler
