+++ server/__init__.py
@@ -1,58 +1,1 @@
-import eventlet
-from flask import Flask
-from database import db
-from flask_migrate import Migrate
-from flask_bcrypt import Bcrypt
-from flask_cors import CORS
-from api.sockets import SocketIO, attach_events
-
-
-migrate = Migrate()
-cors = CORS()
-flask_bcrypt = Bcrypt()
-socketio = SocketIO()
-
-
-def create_app():
-    app = Flask(__name__)
-    app.config.from_object('config.DevelopmentConfig')
-
-    db.init_app(app)
-    migrate.init_app(app, db)
-    cors.init_app(app)
-    flask_bcrypt.init_app(app)
-    socketio.init_app(app)
-    # socketio.run(app)
-    # attach_events(socketio)
-
-    with app.app_context():
-
-        # Register Routes
-        from api.ping_handler import ping_handler
-        from api.home_handler import home_handler
-        from api.auth_handler import auth_handler
-        from api.product_handler import product_handler
-        from api.list_handler import list_handler
-        from api.scrape_handler import scrape_handler
-        from api.list_to_product_handler import list_to_product_handler
-        from api.prices_handler import prices_handler 
-        from api.profile_handler import profile_handler 
-        from api.follower_handler import follower_handler
-        
-        
-
-        # Register Blueprints
-        app.register_blueprint(home_handler)
-        app.register_blueprint(ping_handler)
-        app.register_blueprint(auth_handler)
-        app.register_blueprint(product_handler)
-        app.register_blueprint(list_handler)
-        app.register_blueprint(scrape_handler) 
-        app.register_blueprint(list_to_product_handler) 
-        app.register_blueprint(prices_handler) 
-        app.register_blueprint(profile_handler)
-        app.register_blueprint(follower_handler) 
-
-        db.create_all()
-        return app
