+++ server/bcrypt.py
@@ -1,0 +1,3 @@
+from flask_bcrypt import Bcrypt
+
+bcrypt = Bcrypt()
